# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Server for platform telegram bot [duirole]",
        "title": "Duirole platform API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/v1/test1": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PlatformService"
                ],
                "summary": "Test handler",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/{app_key}/temp_token": {
            "get": {
                "description": "Получение временного токена для бота по ключу. Жизнь токена ограничена 24 часами.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PlatformService"
                ],
                "summary": "Получение временного токена",
                "parameters": [
                    {
                        "type": "string",
                        "description": "key bot application",
                        "name": "app_key",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors_handler.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "errors_handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "localized_error": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}